Refactor project assistant modules for improved organization and clarity
- Moved ConversationProjectManager and related classes to conversation_project_link.py.
- Introduced project_notifications.py for handling project notifications and UI refresh functionality.
- Created project_storage_models.py to define storage-related data models, including ConversationRole.
- Updated imports across various modules to reflect the new structure.
- Removed unused context_types.py and consolidated relevant functionality into existing modules.
- Adjusted tests to accommodate the refactored structure and ensure proper functionality.
