+++ assistants/project-assistant/assistant/project_manager.py
@@ -39,13 +39,10 @@ from .project_data import (
     RequestStatus,
     SuccessCriterion,
 )
-from .project_storage import (
-    ConversationProjectManager,
-    ConversationRole,
-    ProjectNotifier,
-    ProjectStorage,
-    ProjectStorageManager,
-)
+from .conversation_project_link import ConversationProjectManager
+from .project_notifications import ProjectNotifier
+from .project_storage import ProjectStorage, ProjectStorageManager
+from .project_storage_models import ConversationRole
 from .utils import get_current_user, require_current_user
