+++ assistants/project-assistant/tests/test_project_tools.py
@@ -8,7 +8,7 @@ from unittest.mock import AsyncMock, MagicMock
 import openai_client
 import pytest
 from assistant.project_analysis import detect_information_request_needs
-from assistant.project_storage import ConversationRole
+from assistant.project_storage_models import ConversationRole
 from assistant.project_tools import ProjectTools
 from semantic_workbench_assistant.assistant_app import ConversationContext
